Make SignIn form reusable
