+++ packages/design-system/package.json
@@ -43,6 +43,7 @@
     "./components/Label": "./dist/components/Form/Label.js",
     "./components/Separator": "./dist/components/Separator/index.js",
     "./molecules/AsideBrandLayout": "./dist/molecules/AsideBrandLayout/index.js",
+    "./modules/auth/SignInForm": "./dist//modules/auth/SignInForm/index.js",
     "./forms/Field": "./dist/forms/Field/index.js",
     "./forms/useForm": "./dist/forms/useForm/index.js",
     "./main.css": "./dist/main.css"
