+++ packages/design-system/src/modules/auth/SignInForm/EmailPasswordForm/index.tsx
@@ -1,0 +1,1 @@
+export * from './EmailPasswordForm'
