+++ packages/design-system/src/modules/auth/SignInForm/index.tsx
@@ -1,0 +1,1 @@
+export * from './SignInForm'
